refactor: constants 파일에 있는 상수/타입들 파일 분리
